+++ src/main/java/be/baur/sdt/serialization/SDTParser.java
@@ -6,7 +6,6 @@ import java.util.Arrays;
 import java.util.List;
 
 import be.baur.sda.Node;
-import be.baur.sda.NodeSet;
 import be.baur.sda.SDA;
 import be.baur.sdt.Transform;
 import be.baur.sdt.statements.ChooseStatement;
@@ -73,7 +72,10 @@ public final class SDTParser implements Parser {
 	private static final String ATTRIBUTE_NOT_SINGULAR = "attribute '%s' can occur only once";
 	private static final String ATTRIBUTE_REQUIRES_VALUE = "attribute '%s' requires a value";
 	private static final String ATTRIBUTE_UNKNOWN = "attribute '%s' is unknown";
-
+	
+	private static final String NODE_NAME_INVALID = "node name '%s' is invalid";
+	private static final String VARIABLE_NAME_INVALID = "variable name '%s' is invalid";
+	private static final String PARAMETER_REDECLARED = "parameter '%s' cannot be redeclared";
 	
 	@Override
 	public Transform parse(Reader input) throws IOException, ParseException, java.text.ParseException {
@@ -101,13 +103,13 @@ public final class SDTParser implements Parser {
 		if (!sdt.getValue().isEmpty())
 			throw new ParseException(sdt, String.format(STATEMENT_REQUIRES_NO_VALUE, sdt.getName()));
 
-		if (! sdt.isComplex()) // a transform should have a compound statement, even if empty
+		if (sdt.isLeaf()) // a transform should have a compound statement, even if empty
 			throw new ParseException(sdt, String.format(STATEMENT_INCOMPLETE, Transform.TAG));
 
 		checkAttributes(sdt, null); // no attributes allowed in the transform root
 
 		Transform transform = new Transform();
-		for (Node node : sdt.getNodes().find(n -> n.isComplex()))
+		for (Node node : sdt.find(n -> ! n.isLeaf()))
 			transform.add(parseStatement(node)); // parse and add child statements
 
 		return transform;
@@ -187,26 +189,31 @@ public final class SDTParser implements Parser {
 		/*
 		 * A valid variable/param statement has a non-empty value with a variable name,
 		 * a single SELECT attribute containing an XPath expression, and no other
-		 * statements or attributes.
+		 * statements or attributes. Also, parameters must be declared globally (in the
+		 * transform node) and not more than once.
 		 */
-		String value = sdt.getValue();
-		if (value.isEmpty())
+		String varname = sdt.getValue();
+		
+		boolean isParam = sdt.getName().equals(Statements.PARAM.tag);	
+		if ( isParam ) {
+			if (! sdt.getParent().getName().equals(Transform.TAG)) // parent cannot be null
+				throw new ParseException(sdt, String.format(STATEMENT_NOT_ALLOWED, sdt.getName()));	
+			if (sdt.getParent().find(n -> n.getName().equals(Statements.PARAM.tag) && n.getValue().equals(varname)).size() > 1)
+				throw new ParseException(sdt, String.format(PARAMETER_REDECLARED, varname));		
+		}
+		
+		if (varname.isEmpty())
 			throw new ParseException(sdt, String.format(STATEMENT_REQUIRES_VARIABLE, sdt.getName()));
-
+		if (! VariableStatement.isVarName(varname))
+			throw new ParseException(sdt, String.format(VARIABLE_NAME_INVALID, varname));
+		
 		checkStatements(sdt, null); // no sub-statements allowed
 		checkAttributes(sdt, Arrays.asList(Attribute.SELECT));
 		Node select = getAttribute(sdt, Attribute.SELECT, true);
 
-		VariableStatement stat;
-		try {
-			stat = sdt.getName().equals(Statements.VARIABLE.tag) 
-				? new VariableStatement(value, xpathFromNode(select))
-				: new ParamStatement(value, xpathFromNode(select));
-		}
-		catch (Exception e) {
-			throw new ParseException(sdt, e);
-		}
-		return stat;
+		return isParam 
+			? new ParamStatement(varname, xpathFromNode(select))
+			: new VariableStatement(varname, xpathFromNode(select));
 	}
 
 
@@ -225,7 +232,7 @@ public final class SDTParser implements Parser {
 		checkAttributes(sdt, null);
 
 		ForEachStatement statement = new ForEachStatement(xpathFromNode(sdt));
-		for (Node node : sdt.getNodes()) // parse and add child statements
+		for (Node node : sdt.nodes()) // parse and add child statements
 			statement.add(parseStatement(node));
 
 		return statement;
@@ -247,7 +254,7 @@ public final class SDTParser implements Parser {
 		checkAttributes(sdt, null);
 
 		IfStatement statement = new IfStatement(xpathFromNode(sdt));
-		for (Node node : sdt.getNodes()) // parse and add child statements
+		for (Node node : sdt.nodes()) // parse and add child statements
 			statement.add(parseStatement(node));
 
 		return statement;
@@ -270,8 +277,8 @@ public final class SDTParser implements Parser {
 		checkAttributes(sdt, null);
 
 		ChooseStatement statement = null;
-		int i = 0, last = sdt.getNodes().size();
-		for (Node node : sdt.getNodes()) {
+		int i = 0, last = sdt.nodes().size();
+		for (Node node : sdt.nodes()) {
 
 			++i;
 			Statement substat = parseStatement(node);
@@ -308,7 +315,7 @@ public final class SDTParser implements Parser {
 		checkAttributes(sdt, null);
 
 		WhenStatement statement = new WhenStatement(xpathFromNode(sdt));
-		for (Node node : sdt.getNodes()) // parse and add child statements
+		for (Node node : sdt.nodes()) // parse and add child statements
 			statement.add(parseStatement(node));
 
 		return statement;
@@ -330,7 +337,7 @@ public final class SDTParser implements Parser {
 		checkAttributes(sdt, null);
 
 		OtherwiseStatement statement = new OtherwiseStatement();
-		for (Node node : sdt.getNodes()) // parse and add child statements
+		for (Node node : sdt.nodes()) // parse and add child statements
 			statement.add(parseStatement(node));
 
 		return statement;
@@ -351,20 +358,17 @@ public final class SDTParser implements Parser {
 
 		if (nodename.isEmpty())
 			throw new ParseException(sdt, String.format(STATEMENT_REQUIRES_NODENAME, sdt.getName()));
+		if (! SDA.isName(nodename))
+			throw new ParseException(sdt, String.format(NODE_NAME_INVALID, nodename));
 		
 		checkAttributes(sdt, Arrays.asList(Attribute.VALUE));
 		final Node value = getAttribute(sdt, Attribute.VALUE, false);
 			
-		Statement stat;
-		try {
-			stat = (value == null) ? new NodeStatement(nodename)
-				: new NodeValueStatement(nodename, xpathFromNode(value));
-		}
-		catch (Exception e) {
-			throw new ParseException(sdt, e);
-		}
+		Statement stat = (value == null) 
+			? new NodeStatement(nodename) 
+			: new NodeValueStatement(nodename, xpathFromNode(value));
 
-		for (Node node : sdt.getNodes().find(n -> n.isComplex()))
+		for (Node node : sdt.find(n -> ! n.isLeaf()))
 			stat.add(parseStatement(node)); // parse and add child statements
 
 		return stat;
@@ -423,7 +427,7 @@ public final class SDTParser implements Parser {
 	 */
 	private static void checkStatements(Node sdt, List<Statements> allowed) throws ParseException {
 
-		for (Node node : sdt.getNodes().find(n -> n.isComplex())) {
+		for (Node node : sdt.find(n -> ! n.isLeaf())) {
 
 			Statements statement = Statements.get(node.getName());
 			if (statement == null) // all statements must have a known name tag
@@ -444,7 +448,7 @@ public final class SDTParser implements Parser {
 	 */
 	private static void checkAttributes(Node sdt, List<Attribute> allowed) throws ParseException {
 
-		for (Node node : sdt.getNodes().find(n -> !n.isComplex())) {
+		for (Node node : sdt.find(n -> n.isLeaf())) {
 
 			Attribute attribute = Attribute.get(node.getName());
 			if (attribute == null) // all attributes must have a known name tag
@@ -474,8 +478,9 @@ public final class SDTParser implements Parser {
 	 */
 	private static Node getAttribute(Node sdt, Attribute attribute, Boolean required) throws ParseException {
 
-		NodeSet attributes = sdt.getNodes().find(n -> !n.isComplex()).find(attribute.tag);
-		int size = attributes.size();
+		List<Node> alist = sdt.find(n -> n.isLeaf() && n.getName().equals(attribute.tag));
+		int size = alist.size();
+		
 		if (size == 0) {
 			if (required == null || !required)
 				return null;
@@ -484,7 +489,7 @@ public final class SDTParser implements Parser {
 		if (required == null)
 			throw new ParseException(sdt, String.format(ATTRIBUTE_NOT_ALLOWED, attribute.tag));
 
-		Node node = attributes.get(1);
+		Node node = alist.get(0);
 		if (node.getValue().isEmpty())
 			throw new ParseException(node, String.format(ATTRIBUTE_REQUIRES_VALUE, attribute.tag));
 		if (size > 1)