+++ RELEASE_NOTES.md
@@ -1,9 +1,20 @@
 # Release Notes
 
-More than anything else, this release is a proof of concept.
+## [1.1.0] - 2023-10-04
 
-## [1.0.0] - 2023-07-04
-- Supports basic transformation of SDA content using XPath.
+This is a compatibility release for sda-core 2.1.x with minor changes.
+
+- `Changed` parameters to be global variables only.
+- Added *VariableStatement.isVarName* method.
 
 ## Compatibility
-- Requires at least Java 8, sda-core 2.0.0 and Jaxen 2.0.0.
+
+- Requires at least Java 8, sda-core 2.1.0 and Jaxen 2.0.0.
+
+## Previous releases
+
+### [1.0.0] - 2023-07-04 (requires at least SDA v2.0.0)
+
+More than anything else, this release is a proof of concept.
+
+- Supports basic transformation of SDA content using XPath.